requirements_windows.txt
opencv-python
pillow
google-cloud-storage
python-dotenv
